requirements.txt
oauth2client<4.0.0
cachetools==5.3.3
certifi==2024.2.2
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
Jinja2==3.1.2
numpy==1.24.2
pandas==1.5.3
oauthlib==3.2.2
psycopg2-binary==2.9.9
protobuf<3.20
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.1
ratelimiter==1.2.0.post0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==58.0.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.6

